requirements.txt
torch
torchvision
timm
albumentations
grad-cam
matplotlib
numpy
pillow
scikit-learn
pandas
seaborn